requirements.txt
aiogram==3.14.0
asyncpg==0.30.0
python-dotenv==1.0.1
PyMuPDF==1.24.0
